+++ src/components/commons/PostCard.jsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import { MdFavorite, MdFavoriteBorder, MdChatBubbleOutline, MdOutlinePerson } from 'react-icons/md';
 import AuthContext from '@/contexts/auth/auth.context';
 import { getLikes, getComments, getAuthorName } from '@libs/api/supabase.api';
-import formatDate from '@/utils/format.Date';
+import formatDate from '@/utils/formatDate';
 
 export default function PostCard({ postData: postData }) {
   const [likes, setLikes] = useState([]);
